Updated to handle difference between new VIN added and data just changing for VIN
In the search results log and the display output, made modifications so a user can see when a new VIN was added compared to the prior search , and when the field data for the car with that same VIN changed compared to the prior search  (i.e. same VIN but changes for the ETAs, shipping status, hold status, or any other field) along with a list of the names of the fields that changed. to make it easy to identiry what changed.