Modified how and where the specific user match criteria filter is specified
Moved the actual specific user match criteria  filter code to its own python file that the user can specify the name of in the config file via the new userMatchCriteriaFilterFileName config item.   Thus the user creates files for these filters and then uses the config file to specify which filter is to be used.