Updates so that correct error code is returned by run poetry when  update_vehicles is called
Apparently any return statement in vehicles.update_vehicles returns a non zero error code when workflow  run poetry update_vehicles()  is called which causes the github workflow to fail.  So now we have a second function update_vehicles_and_return_df that searchForVehicles.py can call to additionally get that models inventory dataframe as a returned function value as well like it is was expecting when it called the prior version of update_vehicles().   Didn't realize that this is how that works and incorrectly thought you could hava both a function return value and a vallid exit code