Updates to correct condition that determines if all vehicles found.
Updated the condition that determines if all vehicles found to use the unique VIN number of records found >=  the smallest totalRecords shown for any locale (zipcode) search.  This replaces the condition where the first time the number of records added is 0 because depending on the distribution of vehicles geographically that can become true before are vehicles are found.  Also added hawaii and alaska as zipcodes to search as normally those would appear as pages > 40  of the east,west,central search which the totyota website does not make available to access (toyota only allows the first 40 pages to be accesssed)