Updated query locations again for very high count model inventory ,  updated inventory
Added a few more locations for the very high count inventory  to reduce the number of pages needing to be requested and also be under the max 40 pages you can access for a given location request.
Also added debugging option to output csv for each page of data to help determine location coverage.