Updated to leave in entries where ext color was blank or could not find dealer ID
Decided to leave these in the database. so could still see them. as a user could still find the info useful  (especially if only the dealer ID was missing).
Also corrected check for Options to now be iterable as other test did not work when using an existing parquet file (did not handle an array).  Now works whether using an existing parquet file or getting the inventory from the website.