Added logic so that program continues on from last ponit if terminated abnormally
So if the program terminates abnormally, and it is started up again, it will continue on form the point it was at.  The program writes out results every 50 zip codes checked so at most the program would redo checking the last 50 zip codes again, before continuing with new zip code.
Note that there may be a quicker way to get all the toyota dealers but this was the simplest as the only pages we need to request are the https://www.toyota.com/service/tcom/locateDealer/zipCode/<zipcode> pages