Corrected faulty logic that determines added,modified,removed vin entries
The logic used before did not do an index reset of the two dataframes before processing them as the logic used the index label to access entries (.loc).   Use of iloc seemed to be harder since slices don't retain the same iloc and as the dataframe they are sliced from so needed to use .loc indexes.
Also the amount of time the code was taking to determine added, modified, was excessively long (15 minutes or more) when the dataframe got to 20,000 rows or more  so reoptimized the code.  So now it takes about  90 secs to do it.   Also added code to handle the existing outputAddedSearchResultsOnChange value of the config item outputResultsMethod ,  as the code was missing for using it.   To test the added, modified, removed made two mirai parquet test files with one the base and the other with a few added, modified and removed vin entries, and then used the local data option for testing, by copying the test file to the mirai_raw.parquet file before the next pass of the search