Added saving last matches to parquet file so can pick up where left off if terminated
If the programs is terminated and restarted it will use the last matches parquet file as the starting last matches data frame from which to determine added, modified, deleted  entries.  The name of the parquet is the config item resultsFileName  value plus .lastUserMatches.parquet so it is associated with the results file name.