Added useLocalInventoryFile as config item,  added main to vehicles.py
Added useLocalInventoryFile config file item.    So can now run the search program using the local data file instead of querying website.  This opens up several options for running a series of searches for a given model in sequence off a single inventory file for that model to reduce website inventory requests.  Updated descriptions in invocations, and the example config file related to this.