Updates to add Status info file for each model gotten to indicate if get worked and how many vehicles were missing
This  status info file for each model indicated if the get worked (possibly with missing vehicles),  how many vehicles were found, how many vehicles were missing, and the date the get occurred at.  On the google drive this can be used to determine when the model .csv and parquet files were actually created and the degree to which they are valid (google always set the file date to the current upload date and time).  This info is also used by the searchForVehicles.py program to determine if the data is good enough to run its matching criteria on or not so we don't run that and the differences determiner if the file is not good enough.