Corrected logic that determines how many pages/records to get
Stupid mistake that was terminating the get prematurely.   Basically we get the smallest number of total  pages and smallest number of total records info returned by any location for the given page number we are on and that becomes the max pages and records we are looking for that determine if we are done or not on that pass.